fix： replace gettables
